@@ -1,26 +1,26 @@
 declare module '*.scss' {
-    const content: Record<string, string>;
-    export default content;
+  const content: Record<string, string>;
+  export default content;
 }
 
-declare module '\*.svg' {
-    import React = require('react');
-    const ReactComponent: React.FC<React.SVGProps<SVGSVGElement>>;
-    export default ReactComponent;
+declare module '*.svg' {
+  import React = require('react')
+  const ReactComponent: React.FC<React.SVGProps<SVGSVGElement>>;
+  export default ReactComponent;
 }
-declare module '\*.jpg' {
-    const content: string;
-    export default content;
-}
-
-declare module '\*.png' {
-    const content: string;
-    export default content;
+declare module '*.jpg' {
+  const content: string;
+  export default content;
 }
 
-declare module '\*.json' {
-    const content: string;
-    export default content;
+declare module '*.png' {
+  const content: string;
+  export default content;
+}
+
+declare module '*.json' {
+  const content: string;
+  export default content;
 }
 
 declare const __DEV__: boolean;
