@@ -1,13 +1,13 @@
-import React, {Suspense} from 'react';
-import {Route, Routes} from "react-router-dom";
-import {RouteConfig} from "shared/config/routeConfig/routeConfig";
+import { Suspense } from 'react';
+import { Route, Routes } from 'react-router-dom';
+import { RouteConfig } from 'shared/config/routeConfig/routeConfig';
 
 export const AppRouter = () => {
   return (
     <Suspense fallback={<div>loading</div>}>
       <Routes>
         {
-          Object.values(RouteConfig).map(({element, path}) => (
+          Object.values(RouteConfig).map(({ element, path }) => (
             <Route
               key={path}
               element={element}
