@@ -1,16 +1,16 @@
-import {render} from "react-dom";
+import { render } from 'react-dom';
 import './app/styles/index.scss';
-import {BrowserRouter} from "react-router-dom";
-import {ThemeProvider} from "app/providers/ThemeProvider";
-import App from "app/App";
+import { BrowserRouter } from 'react-router-dom';
+import { ThemeProvider } from 'app/providers/ThemeProvider';
+import App from 'app/App';
 
-import "./shared/config/i18n/i18n";
+import './shared/config/i18n/i18n';
 
-render (
-    <BrowserRouter>
-        <ThemeProvider>
-            <App/>
-        </ThemeProvider>
-    </BrowserRouter>,
-    document.getElementById('root')
-)
+render(
+  <BrowserRouter>
+    <ThemeProvider>
+      <App/>
+    </ThemeProvider>
+  </BrowserRouter>,
+  document.getElementById('root'),
+);
