@@ -1,3 +1,3 @@
-import {lazy} from "react";
+import { lazy } from 'react';
 
-export const AboutPageAsync = lazy(() => import('./AboutPage'));
+export const AboutPageAsync = lazy(async () => await import('./AboutPage'));
