@@ -1,14 +1,14 @@
 import React from 'react';
-import {useTranslation} from "react-i18next";
+import { useTranslation } from 'react-i18next';
 
 const AboutPage = () => {
-    const {t} = useTranslation('about');
+  const { t } = useTranslation('about');
 
-    return (
-        <div>
-            {t('About page')}
-        </div>
-    );
+  return (
+    <div>
+      {t('About page')}
+    </div>
+  );
 };
 
 export default AboutPage;
