@@ -1,3 +1,3 @@
-import {lazy} from "react";
+import { lazy } from 'react';
 
-export const MainPageAsync = lazy(() => import('./MainPage'));
+export const MainPageAsync = lazy(async () => await import('./MainPage'));
