@@ -1,14 +1,13 @@
-import React from 'react';
-import {useTranslation} from "react-i18next";
+import { useTranslation } from 'react-i18next';
 
 const MainPage = () => {
-    const {t} = useTranslation('main');
+  const { t } = useTranslation('main');
 
-    return (
-        <div>
-            {t('Main page')}
-        </div>
-    );
+  return (
+    <div>
+      {t('Main page')}
+    </div>
+  );
 };
 
 export default MainPage;
