@@ -1,5 +1,5 @@
 import i18next from 'i18next';
-import {initReactI18next} from "react-i18next";
+import { initReactI18next } from 'react-i18next';
 
 import Backend from 'i18next-http-backend';
 import LanguageDetector from 'i18next-browser-languagedetector';
@@ -9,9 +9,9 @@ i18next
   .use(LanguageDetector)
   .use(initReactI18next)
   .init({
-    fallbackLng: 'ru',
-    debug: __DEV__,
-    interpolation: {
-      escapeValue: false,
-    }
-  })
+    'fallbackLng': 'ru',
+    'debug': __DEV__,
+    'interpolation': {
+      'escapeValue': false,
+    },
+  });
