@@ -1,24 +1,24 @@
 import React from 'react';
-import {RouteProps} from "react-router-dom";
-import {AboutPage, MainPage} from "pages";
+import { type RouteProps } from 'react-router-dom';
+import { AboutPage, MainPage } from 'pages';
 
 export enum AppRoutes {
-    MAIN = 'main',
-    ABOUT = 'about',
+  MAIN = 'main',
+  ABOUT = 'about',
 }
 
-export const RoutePath: Record<AppRoutes, string> =  {
-    [AppRoutes.MAIN]: '/',
-    [AppRoutes.ABOUT]: '/about',
+export const RoutePath: Record<AppRoutes, string> = {
+  [AppRoutes.MAIN]: '/',
+  [AppRoutes.ABOUT]: '/about',
 };
 
 export const RouteConfig: Record<AppRoutes, RouteProps> = {
-    [AppRoutes.MAIN]: {
-        path: RoutePath[AppRoutes.MAIN],
-        element: <MainPage />,
-    },
-    [AppRoutes.ABOUT]: {
-        path: RoutePath[AppRoutes.ABOUT],
-        element: <AboutPage />,
-    },
-}
+  [AppRoutes.MAIN]: {
+    'path': RoutePath[AppRoutes.MAIN],
+    'element': <MainPage />,
+  },
+  [AppRoutes.ABOUT]: {
+    'path': RoutePath[AppRoutes.ABOUT],
+    'element': <AboutPage />,
+  },
+};
