@@ -1,12 +1,11 @@
-type Mods = Record<string, boolean> | string;
+type Mods = Record<string, boolean> | string
 
-
-export const classNames = (cls: string, mods: Mods = {}, additional: string[] = []): string => {
-    return [
-        cls,
-        ...additional.filter(Boolean),
-        ...Object.entries(mods)
-            .filter(([, value]) => Boolean(value))
-            .map(([className]) => className)
-    ].join(' ');
-}
+export const classNames = (cls: string, mods: Mods = {}, additional: Array<string | undefined> = []): string => {
+  return [
+    cls,
+    ...Object.entries(mods)
+      .filter(([, value]) => Boolean(value))
+      .map(([className]) => className),
+    ...additional.filter(Boolean),
+  ].join(' ');
+};
